Add first integration test for AgentData
This gives at least a basic guarantee against regressions when making changes
to code.

Authorization is set up in a way, that when new tests are added, then the
should be first run with the `TOKEN` environment variable set. VCR will then
record the requests and responses, masking the authorization header, and
subsequent test runs don't need the `TOKEN` environment variable to be set
anymore.