Add custom code for menu that's using pseudo-classes
COB-614

There's a client that has a navigation menu which relies on `:hover`
pseudo-class for showing menu dropdowns. It's not possible to
trigger pseudo-classes on elements programmatically, meaning that
cobra can't handle them.

This script adds a parallel method for showing menu dropdowns. Namely,
menu items' styles are changed based on user's mouse movements.
E.g. when user's mouse enters a menu item, then dropdown's styles
are set to same values as it's done via CSS. This is something that
cobra detects, and is able to mirror to the operator.

The necessary event listeners are added when an engagement starts,
and removed when engagement ends.

This is definitely not a universal solution. The targeted CSS/HTML
package is [Sky Mega Menu][0]. More specifically, the script addresses
the menu structure and setup for a [certain client][1]. For all other
cases, it's more of like an example of how we *might* handle pseudo-classes.

Disclaimers:
1. This script is an ugly hack. Do not try this at home.
2. The script relies heavily on menu structure. It may or may not
   work for other menus using Sky Mega Menu. It may or not work
   after client has altered their menu.
3. The script handles opening and closing dropdowns, but not
   highlighting menu items on hover. Although rather easy to
   implement, it's not a priority and I did not have the time.
4. Although having tested it thoroughly, there's a risk of this
   script breaking menu functionality or, especially during
   engagements.

[0]: http://voky.com.ua/showcase/sky-mega-menu/
[1]: https://admin.salemove.com/sites/bb33caa8-b8da-4a47-a76a-6328269e0315